+++ src/tx-decorator/convert-buy.js
@@ -1,0 +1,3 @@
+import decorateBuyTxParams from './convert-sell';
+
+export default decorateBuyTxParams;