+++ src/tx-decorator/convert-sell-all.js
@@ -1,0 +1,3 @@
+import decorateSellAllTxParams from './convert-sell';
+
+export default decorateSellAllTxParams;