+++ src/tx-decorator/stake-delegate.js
@@ -1,0 +1,3 @@
+import decorateDelegateTxParams from './candidacy-declare';
+
+export default decorateDelegateTxParams;