+++ test/link.test.js
@@ -1,89 +1,119 @@
-import {TxDataRedeemCheck} from 'minterjs-tx';
-import {prepareLink, decodeLink, SendTxParams, RedeemCheckTxParams} from '~/src';
-import getTxData from '~/src/tx-data';
+import {TX_TYPE} from 'minterjs-tx';
+import {prepareLink, decodeLink} from '~/src';
+import RedeemCheckTxData from '~/src/tx-data/redeem-check';
+import {VALID_CHECK} from '~/test/check.test';
 
-const TX_PARAMS_DATA_SEND = {
-    address: 'Mx7633980c000139dd3bd24a3f54e06474fa941e16',
-    amount: 10,
-    coinSymbol: 'MNT',
-    feeCoinSymbol: 'ASD',
+const TX_PARAMS_SEND = {
+    type: TX_TYPE.SEND,
+    data: {
+        coin: 'MNT',
+        to: 'Mx7633980c000139dd3bd24a3f54e06474fa941e16',
+        value: '10',
+    },
+    gasCoin: 'ASD',
     payload: 'custom message',
 };
 const LINK_SEND = 'https://bip.to/tx?d=f84801aae98a4d4e5400000000000000947633980c000139dd3bd24a3f54e06474fa941e16888ac7230489e800008e637573746f6d206d65737361676580808a41534400000000000000';
-const TX_PARAMS_DATA_CHECK = {
-    check: 'Mcf8ab3101830f423f8a4d4e5400000000000000888ac7230489e800008a4d4e5400000000000000b841f69950a210196529f47df938f7af84958cdb336daf304616c37ef8bebca324910910f046e2ff999a7f2ab564bd690c1102ab65a20e0f27b57a93854339b60837011ba00a07cbf311148a6b62c1d1b34a5e0c2b6931a0547ede8b9dfb37aedff4480622a023ac93f7173ca41499624f06dfdd58c4e65d1279ea526777c194ddb623d57027',
-    password: '123',
-    privateKey: '5fa3a8b186f6cc2d748ee2d8c0eb7a905a7b73de0f2c34c5e7857c3b46f187da',
+
+const TX_PARAMS_MULTISEND = {
+    type: TX_TYPE.MULTISEND,
+    data: {
+        list: [
+            {
+                value: 0.1,
+                coin: 'MNT',
+                to: 'Mxfe60014a6e9ac91618f5d1cab3fd58cded61ee99',
+            },
+            {
+                value: 0.2,
+                coin: 'MNT',
+                to: 'Mxddab6281766ad86497741ff91b6b48fe85012e3c',
+            },
+        ],
+    },
 };
-const LINK_CHECK = 'https://bip.to/tx?d=f9010509b8f4f8f2b8adf8ab3101830f423f8a4d4e5400000000000000888ac7230489e800008a4d4e5400000000000000b841f69950a210196529f47df938f7af84958cdb336daf304616c37ef8bebca324910910f046e2ff999a7f2ab564bd690c1102ab65a20e0f27b57a93854339b60837011ba00a07cbf311148a6b62c1d1b34a5e0c2b6931a0547ede8b9dfb37aedff4480622a023ac93f7173ca41499624f06dfdd58c4e65d1279ea526777c194ddb623d57027b8417adcf6a62a66b177b266c767c5ebd906651fb66269401a8c66d053574dc29c67296b93af2e276fbdf5f606a98419ae69191450f67a2d273ee6c5d3016773c16d018080018a4d4e5400000000000000';
-const LINK_CHECK_PASSWORD = 'https://bip.to/tx?d=f8c309b8b2f8b0b8adf8ab3101830f423f8a4d4e5400000000000000888ac7230489e800008a4d4e5400000000000000b841f69950a210196529f47df938f7af84958cdb336daf304616c37ef8bebca324910910f046e2ff999a7f2ab564bd690c1102ab65a20e0f27b57a93854339b60837011ba00a07cbf311148a6b62c1d1b34a5e0c2b6931a0547ede8b9dfb37aedff4480622a023ac93f7173ca41499624f06dfdd58c4e65d1279ea526777c194ddb623d57027808080018a4d4e5400000000000000&p=83313233';
+const LINK_MULTISEND = 'https://bip.to/tx?d=f85f0db858f856f854e98a4d4e540000000000000094fe60014a6e9ac91618f5d1cab3fd58cded61ee9988016345785d8a0000e98a4d4e540000000000000094ddab6281766ad86497741ff91b6b48fe85012e3c8802c68af0bb14000080808080';
+
+const TX_PARAMS_CHECK = {
+    type: TX_TYPE.REDEEM_CHECK,
+    data: {
+        check: VALID_CHECK,
+        password: 'pass',
+        privateKey: '5fa3a8b186f6cc2d748ee2d8c0eb7a905a7b73de0f2c34c5e7857c3b46f187da',
+    },
+};
+const LINK_CHECK = 'https://bip.to/tx?d=f8fb09b8f4f8f2b8adf8ab3101830f423f8a4d4e5400000000000000888ac7230489e800008a4d4e5400000000000000b841f69950a210196529f47df938f7af84958cdb336daf304616c37ef8bebca324910910f046e2ff999a7f2ab564bd690c1102ab65a20e0f27b57a93854339b60837011ba00a07cbf311148a6b62c1d1b34a5e0c2b6931a0547ede8b9dfb37aedff4480622a023ac93f7173ca41499624f06dfdd58c4e65d1279ea526777c194ddb623d57027b8410497ea588f0fc2bd448de76d03a74cf371269e10ac1a02765fb5fa37c29f67e0348fb3faacd3370b8809401e7d562d8943f3642ce96667188d3c344e8e5bff6d0180808080';
+const LINK_CHECK_PASSWORD = 'https://bip.to/tx?d=f8b909b8b2f8b0b8adf8ab3101830f423f8a4d4e5400000000000000888ac7230489e800008a4d4e5400000000000000b841f69950a210196529f47df938f7af84958cdb336daf304616c37ef8bebca324910910f046e2ff999a7f2ab564bd690c1102ab65a20e0f27b57a93854339b60837011ba00a07cbf311148a6b62c1d1b34a5e0c2b6931a0547ede8b9dfb37aedff4480622a023ac93f7173ca41499624f06dfdd58c4e65d1279ea526777c194ddb623d570278080808080&p=8470617373';
 
 
 describe('prepareLink()', () => {
     test('should work', () => {
-        const link = prepareLink(new SendTxParams(TX_PARAMS_DATA_SEND));
+        const link = prepareLink(TX_PARAMS_SEND);
         expect(link).toEqual(LINK_SEND);
+    });
+
+    describe('multisend', () => {
+        test('should work', () => {
+            const link = prepareLink(TX_PARAMS_MULTISEND);
+            expect(link).toEqual(LINK_MULTISEND);
+        });
     });
 
     describe('check', () => {
         test('should work with proof', () => {
-            const redeemCheckTxParams = new RedeemCheckTxParams(TX_PARAMS_DATA_CHECK);
-            const link = prepareLink(redeemCheckTxParams);
+            const data = new RedeemCheckTxData(TX_PARAMS_CHECK.data);
+            const link = prepareLink({...TX_PARAMS_CHECK, data});
             expect(link).toEqual(LINK_CHECK);
         });
 
         test('should work with password', () => {
-            const redeemCheckTxParams = new RedeemCheckTxParams(TX_PARAMS_DATA_CHECK);
-            const redeemCheckTxParamsWithoutProof = removeProofFromData(redeemCheckTxParams);
             const link = prepareLink({
-                ...redeemCheckTxParamsWithoutProof,
-                password: '123',
+                type: TX_PARAMS_CHECK.type,
+                data: {check: TX_PARAMS_CHECK.data.check, password: TX_PARAMS_CHECK.data.password},
+                password: TX_PARAMS_CHECK.data.password,
             });
             expect(link).toEqual(LINK_CHECK_PASSWORD);
         });
     });
+
+    describe('host', () => {
+        const testnetLink = LINK_SEND.replace('https://bip.to', 'https://testnet.bip.to');
+        test('should work with custom host', () => {
+            const link = prepareLink(TX_PARAMS_SEND, 'https://testnet.bip.to');
+            expect(link).toEqual(testnetLink);
+        });
+        test('should work with ending slash', () => {
+            const link = prepareLink(TX_PARAMS_SEND, 'https://testnet.bip.to/');
+            expect(link).toEqual(testnetLink);
+        });
+        test('should work without scheme', () => {
+            const link = prepareLink(TX_PARAMS_SEND, 'testnet.bip.to');
+            expect(link).toEqual(testnetLink);
+        });
+    });
 });
-
-function removeProofFromData(txParams) {
-    const redeemCheckTxData = new TxDataRedeemCheck(txParams.txData);
-    // delete proof from data
-    redeemCheckTxData.proof = Buffer.from([]);
-
-    return {
-        ...txParams,
-        txData: redeemCheckTxData.serialize(),
-    };
-}
 
 
 describe('decodeLink()', () => {
     test('should work', () => {
         const txParams = decodeLink(LINK_SEND);
-        const sendTxParams = new SendTxParams(TX_PARAMS_DATA_SEND);
-        sendTxParams.txData = getTxData(sendTxParams.txType).fromRlp(sendTxParams.txData).fields;
-        expect(txParams).toEqual(sendTxParams);
+        expect(txParams).toEqual(TX_PARAMS_SEND);
     });
 
     describe('check', () => {
         test('should work with proof', () => {
             const txParams = decodeLink(LINK_CHECK);
-            const validTxParams = new RedeemCheckTxParams(TX_PARAMS_DATA_CHECK);
-            const txData = getTxData(validTxParams.txType).fromRlp(validTxParams.txData);
-            validTxParams.txData = {check: txData.check, proof: txData.proof};
-            // clean valid params
-            delete validTxParams.privateKey;
+            // add proof
+            const validTxParams = {type: TX_PARAMS_CHECK.type, data: new RedeemCheckTxData(TX_PARAMS_CHECK.data).fields};
             // ensure string payload
             validTxParams.payload = validTxParams.payload || '';
             expect(txParams).toEqual(validTxParams);
         });
 
         test('should work with password', () => {
-            const txParams = decodeLink(LINK_CHECK_PASSWORD, TX_PARAMS_DATA_CHECK.privateKey);
-            const validTxParams = new RedeemCheckTxParams(TX_PARAMS_DATA_CHECK);
-            const txData = getTxData(validTxParams.txType).fromRlp(validTxParams.txData);
-            validTxParams.txData = {check: txData.check, proof: txData.proof};
-            // ensure string private key
-            validTxParams.privateKey = validTxParams.privateKey.toString('hex');
+            const txParams = decodeLink(LINK_CHECK_PASSWORD, TX_PARAMS_CHECK.data.privateKey);
+            // add proof
+            const validTxParams = {type: TX_PARAMS_CHECK.type, data: new RedeemCheckTxData(TX_PARAMS_CHECK.data).fields};
             // ensure string payload
             validTxParams.payload = validTxParams.payload || '';
             expect(txParams).toEqual(validTxParams);