+++ CHANGELOG.md
@@ -17,6 +17,22 @@
 - `redeemCheck` params will check if passed `feeCoinSymbol` is equal to `gasCoin` from check, may be omitted, `gasCoin` from check will be used
 - add `maxSupply` field to `CreateCoinTxParams`
 
+## 0.27.0 - 2020-01-20
+- **BREAKING** Rename: all `rawCheck` fields to `check`
+- Update: allow use `RedeemCheckTxData` without `proof` and `privateKey` it will lead to empty proof
+
+## 0.26.1 - 2020-01-20
+- Add: option for `prepareLink` to allow setting custom hostname
+
+## 0.26.0 - 2020-01-15
+- **BREAKING** Change: `decodeLink` now returns `data` instead of `txData` and `type` instead of `txType` fields
+- **BREAKING** Remove: `privateKey` fields from `decodeLink` result
+- **BREAKINK** Remove: `issueCheck` and `decodeCheck` from `Minter` instance
+- Add: `postTx`, `prepareSignedTx`, and `prepareLink` now accepts `data` and `type` fields, `txData` and `txType` are deprecated
+- Add: `postTx` and `prepareSignedTx` now also accept data object for `data` field, not only Buffer
+- Add: `postTx` and `prepareSignedTx` now decorate txParams as TxParamsConstructors previously do
+- Add: reexport `TX_TYPE` from `minterjs-tx`
+
 ## 0.25.0 - 2019-12-25
 - **BREAKING** Change: `decodeLink` now decode `txData` too
 - Add: `getTxData` method which return TxData constructor by txType